+++ examples/deprecations.js
