+++ examples/example-custom-format.js
@@ -3,7 +3,7 @@
 const http = require('http')
 const server = http.createServer(handle)
 
-const logger = require('./')({
+const logger = require('..')({
   quietReqLogger: true,
   transport: {
     target: 'pino-http-print',
